requirements.txt
pyautogui>=0.9.54
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
spotipy>=2.23.0
python-dotenv>=1.0.0